+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
 zoom:
 
 * new: Add the default value to the caller parameter of the reset() method of ZoomableState and ZoomableEngine
+* new: ZoomableState and ZoomableEngine have added the keepTransformWhenSameAspectRatioContentSizeChanged parameter. After turning on, you can keep the transformation state when switching pictures with the same aspect ratio. [#58](https://github.com/panpf/zoomimage/issues/58)
 
 ## 1.1.2